Restructured the rod node and how it stores data. Introduced the RodGroup class that manages all the rod data and moved a lot of stuff out of Beaker that should have been in the rod node.
NOTE: Nurbs input currently does not work.
I also fixed bugs where rods were not reinitialised if you changed an attribute on the start frame. There's a lot of restructuring of the rod node and beaker with these changes so there may be bugs I've not noticed. Still working on it.

Refs #11084