Fix tolerance levels in CandidateCollision.cc
COLLISION_EPSILON was set to 1e-6, which was small enough to miss
collisions.  Changing to 1e-4 fixed the problem.  It's possible
this difference is due to switching between meters and centimeters.

In addition, the code was using the same tolerance value for
comparisons against values with units of both length and length^2.
Most of the length^2 comparisons were for normalization, so we
replace those with checks for exact zeros.

Refs #16438