Fix line search and change defaults back to 1e-8a.  Refs #17835
When I wrote the line search I introduced a bug which caused the
rhs to be computed incorrectly after the first iteration.  It should
be working correctly now.  In addition, the line search now accepts
any change which doesn't increase the residual by more than a factor
of 10.  This removes the Newton explosions without changing behavior
in the common case.